Invoke TProcessorEventHandler::onInteractionTerminate on tile destruction
Summary: Guarantee that `TProcessorEventHandler::onInteractionTerminate()` is called when the interaction is destroyed. Otherwise the callback is invoked when the client is disconnected, which is not necessarily after all requests have been processed. For instance, a client sending a one-way request may disconnect immediately after sending said request (because a response is not waited for), triggering an `onInteractionTerminate()` event before other read/write callbacks.

Reviewed By: praihan

Differential Revision: D58428575

fbshipit-source-id: 6c381655658a5ba7956d206d3f54fd144d0b5efa