Initial implementation of MutableUnion
Summary:
Currently only supports assignment on construction and access through the reserved "current field" and "current value" attributes.

Note that, as of now, these attributes have different names than the immutable version, in order to ensure that they are reserved (see description of previous commit). This may change in the future.

Reviewed By: yoney

Differential Revision: D58528231

fbshipit-source-id: d1ff90806a433b84d947bb935606d04480d7631f