fix rocket to not panic on close when no messages have been sent
Summary:
fix rocket to not panic on close when no messages have been sent
This was initially found in D58828503

"""
we discovered a bug affecting Netstate (with Rocket) on NetOS hosts.

(But the bug is not limited to NetOS hosts - it just happens to manifests itself on NetOS hosts, because they differ from regular hosts.)

The bug causes a panic. It occurs when a client is opened and then immediately closed without making any RPC calls.
The protocol field remains nil and we get a nil dereference upon Close().
""" - echistyakov

Reviewed By: podtserkovskiy

Differential Revision: D58864210

fbshipit-source-id: 3533d808e5b641c802f6eb9acb5c6478983a346a