Fix conflict between doc(hidden) and doc(inline)
Summary:
The Rust Thrift services crates are affected by a rustdoc bug which makes `doc(hidden)` ineffective when `doc(inline)` is used on a public re-export of a dependency.

Working around this is important for {D58915758} so that the services crate for Thrift targets containing 0 services can be correctly identified as empty and excluded from rustdoc search.

Reviewed By: diliop

Differential Revision: D58915759

fbshipit-source-id: 6c43bcb59898d64a70b90bbc5c4e9bee6ef57d7a