Skip unnecessary chain of MeteredExecutors in SFQThreadManager
Summary:
When `SFQThreadManager` is constructed from a non-`ThreadManager` executor, the base class `ThreadManagerExecutorAdapter` constructs a chain of `MeteredExecutor`s to serve the `UPSTREAM` source, and that becomes in turn the wrapped executor for the per-tenant queues.

This is inefficient, all the prioritization work is already done by `SFQThreadManager`'s `MeteredExecutor`s, and further chaining the submission adds latency to dequeue operations, which can limit concurrency (see for example benchmarks in D58999318).

In this case, we can bypass the internal `MeteredExecutor`s by using `Source::INTERNAL` instead of `UPSTREAM`.

While at it, support configuring `MeteredExecutor`'s `maxInQueue`, introduced in D58999318.

Differential Revision: D58999827

fbshipit-source-id: a64e1082ba1d0dde7b86e557eb2bb05527c9141b