replace conn timeout with manual timeout
Summary:
See the unit test in the testplan that reproduces the error

We have replaced connTimeout with a normal net.Conn to avoid timeouts being trigger by keep alives and full duplex reading.
We then reintroduced the timeout using a single channel in readPayload.

Unfortunately this made other tests fail.
It seems that a Deadline is set inside the rsocket library, but if we sleep long enough, which is necessary in the case of the unit test testRequestReplyTimeout2, but also happens as part of TestParallelRequestReply just because of how long the test runs, then we lose the connection.
`p.conn.SetDeadline(time.Time{})` fixes this by resetting the deadline to not trigger at the writing of every message.

Reviewed By: podtserkovskiy

Differential Revision: D59076844

fbshipit-source-id: 3ff4dfc3eeff3d8cfc7ae5f06f9408ce5fc9ad40