Fix exception attribution for ServiceInterceptor
Summary:
The current implementation uses the index of the  exception in the list of collected exceptions to find the `ServiceInterceptor` to blame.

This is completely wrong! If any of the interceptors in the list do not throw, then the index will be off, and the attribution will be incorrect. See unit tests.

Reviewed By: iahs

Differential Revision: D59294966

fbshipit-source-id: 7345d513f0490ce16408cdd280cebc1e5e5cfc88