Defer schema deserialization
Summary:
The hydrated schema structs can consume a large amount of memory in applications with many thrift files linked in.
To avoid imposing this cost on binaries that won't use the functionality we only store pointers to the serialized data during initialization.
The deserialization takes place on first access instead. (Applications can add an access near the start of main to avoid synchronization cost / blocking).

Reviewed By: thedavekwon

Differential Revision: D59553061

fbshipit-source-id: e6021e995b8e4cc1228f3d217b8d2de76cbf3b1e