Conditionally enable coroutines if supported
fbthrift is still on C++17, so GCC will only enable support for
coroutines if -fcoroutines is passed, which is not the case currently.
This, however, creates confusion when fbthrift is used as a dependency
of a project that's already using C++20 or newer. In such a scenario,
GCC >= 11 will enable coroutine support by default, so generated client
code will be compiled using coroutines and then fail because fbthrift
was built without coroutine support.

So, conditionally enable coroutines if the compiler supports them, like
folly and co. do already.