Propagate function signature change.
Summary:
D60184917 added code that uses `TServerObserver` and committed on Jul 25, 2024 at 2:07 PM.

D60255766 changed the signature of the `TServerObserver::resourcePoolDisabled` and `TServerObserver::resourcePoolEnabled`, and committed on Jul 25, 2024 at 8:35 PM.

Looks like there was a race and the change to the signature in D60255766 did not propagate to the change in D60184917. This (obviously) caused failed tests.

Propagate the signature change.

Reviewed By: yoney

Differential Revision: D60340669

fbshipit-source-id: 912de73fd43c0f305ae8691c48f8eca5660c1ea2