Conditionally enable coroutines if supported (#615)
Summary:
fbthrift is still on C++17, so GCC will only enable support for coroutines if -fcoroutines is passed, which is not the case currently. This, however, creates confusion when fbthrift is used as a dependency of a project that's already using C++20 or newer. In such a scenario, GCC >= 11 will enable coroutine support by default, so generated client code will be compiled using coroutines and then fail because fbthrift was built without coroutine support.

So, conditionally enable coroutines if the compiler supports them, like folly and co. do already.

Pull Request resolved: https://github.com/facebook/fbthrift/pull/615

Reviewed By: Gownta

Differential Revision: D60353408

Pulled By: yfeldblum

fbshipit-source-id: b29d15e91333d2ee4bf65023b9ccbe4697898c2b