Replace link_whole for static init
Summary:
When using static linkage, link_whole is both expensive and unnecessary. Instead, we can name individual symbols for the linker to retain.
This means the symbol needs to have default visibility, which means neither it nor its type can be in an anonymous namespace. Instead, we generate a unique name in the build system and pass that to the Thrift compiler so it can match the linker.

Reviewed By: thedavekwon

Differential Revision: D60303008

fbshipit-source-id: 122cc686cf9b5d6a87ef55253a7f82d8cb7fca3e