Reorganize code to prepare for the changes to Migrate Thrift Python to Resource Pools
Summary:
As I worked on D59727112, which is an incremental step in the change for T176105713 `Migrate Thrift Python to ResourcePools`, I cleaned up code. The review correctly identified this as a clean up that should be in its own diff to make it easier to review the two different types of changes.

This change is effectively a no-op.

SPECIAL NOTE:
server.h and server.cpp contained `PythonAsyncProcessor` and `PythonAsyncProcessorFactory` related code. There was already a `PythonAsyncProcessorFactory.h`. To make things more consistent, this diff
* Renames `server.h` -> `PythonAsyncProcessor.h`
* Splits `server.cpp` into `PythonAsyncProcessor.cpp` and `PythonAsyncProcessorFactory.cpp`

Reviewed By: Filip-F

Differential Revision: D60001767

fbshipit-source-id: a64fe03b826f208d30256e5401c9aac132742cc1