Allow ServiceInterceptor::onResponse to swallow active exception
Summary:
Currently, if `ServiceInterceptor::onResponse` throws and there is an active exception (from either `ServiceInterceptor::onRequest` or user handler code), then the latter "wins" and is returned to the client.

An implementation of `ServiceInterceptor` (such as ACLChecker) might want to block even exceptions being returned to the client.

This diff flips the tie-breaker in favor of `ServiceInterceptor::onResponse`. In addition, `ServiceInterceptor::onResponse` is now provided the active exception as a parameter and thus can selectively decide to throw or not based on its contents.
 ---
Control flow before:
 {F1764035241}
 ---
Control flow after:
 {F1764034373}

Reviewed By: sazonovkirill

Differential Revision: D59988196

fbshipit-source-id: 3c2b881fb669fb6b448308971cccbca3eac8a814