remove concurrency from rocket client
Summary:
remove concurrency from rocket client

This now makes sure that the rocket client passes the SR parallel tests, since Flush blocks until a message is received, this means we cannot close the connection, until the message has been completely sent.

Previously the rocket client could be half way through sending a message and SR would try to close the connection. Tests were passing if we removed the closing logic. This is why we also added an extra test, to make sure we are actually closing the connection and not cheating to pass the parallel test.

Reviewed By: leoleovich

Differential Revision: D60509622

fbshipit-source-id: b4cd4cab0a6aef7a048ea62a8f700d7be1c052ca