Delete cpp.TriviallyRelocatable
Summary:
This warranting approach to marking structs trivially relocatable has proven to be a footgun in the presence of schema evolution.
It is also minimally used - our codebase only has 4 instances, of which 3 are incorrect. As such, this diff simply deletes the functionality.
In the future someone may choose to implement deducing whether a struct should have this property, which would allow the same optimization without the brittleness.

Reviewed By: Mizuchi

Differential Revision: D60600609

fbshipit-source-id: 7f099deed6dfbdbd410ad0c0bc87bf9b6fd944c9