annotate generated fields
Summary:
Second attempt on D60476484 which was reverted as it broke some targets.

The  broken targets depended on this target
```
fbcode//thrift_sync/www/flib/__generated__/IAFEventsMeerkatStep:iaf_events-cpp2-types
```
It failed to build because the generated files were too big for clang to handle (P1522317820). The generated file are big, and these extra comments were the straw that broke the caml's back

We don't add the file, type in the comment anymore, as the indexer can deduce them from the comment on the enclosing class (D61028619)

Note: if needed we could comment only the first field of the list, and have the indexer deduce the rest.

Reviewed By: yoney, bochko

Differential Revision: D60844431

fbshipit-source-id: b469ff2ac2614345c4a2edeb68c74c418e4718a2