Refactor code to prepare for later changes.
Summary:
A later change becomes simpler with this refactor. Rather than include the refactor along with the meat of the change, I chose to make this a separate diff that we can land separately.

In the process, I made the `handlePythonServerCallback...` functions private because there is no reason for them to be public.

Reviewed By: praihan

Differential Revision: D61151507

fbshipit-source-id: 78ca2527b31c9e65c601c9eebed26461853343ef