Access the enable_resource_pools_for_python flag to ensure it is visible in thriftDbg.
Summary:
For the resource pools rollout, we want to use `thriftDbg info` to view the value of the `enable_resource_pools_for_python` flag BEFORE the changes to enable resource pools for python land on the target servers. `thriftDbg info` requires that the execution access the flag at least once before it displays the value of the flag.

This diff allows us to land the changes that add the flag and access it ahead of the configerator change to enable the `enable_resource_pools_for_python` for the pilot service.

Reviewed By: praihan

Differential Revision: D61219842

fbshipit-source-id: 7ad0c35aaccdcf27ddca0452c566f92ca2ee0096