Simplify the implementation related to PythonAsyncProcessorFactory::genericProcessor
Summary:
The current implementation of `PythonAsyncProcessorFactory::genericProcessor` is a function template. Based on the three different RpcKind values that it serves, there is no reason for it to be a function template, which adds to complexity and affects maintainability.

To improve maintainabliity, eliminate the function template and the code that depends on it.

Reviewed By: praihan

Differential Revision: D61351067

fbshipit-source-id: 812d5b6301a5343436fc40c311773aaf1b7a7801