Implement whisker::parser
Summary:
This diff implements the parser for Whisker. It builds on top of the Whisker lexer implemented in {D60944830}. Since Whisker is mostly a superset of mustache, it should be able to correctly parse our existing mustache code.

The structure of the parser code is based on Thrift compiler's own parser. We re-use some machinery: `source`, `source_location`, `source_range`, `diagnostic_engine` etc. However, the equivalent of `parser_actions` is not necessary at the moment because the AST is much simpler than the `t_*` node hierarchy that the Thrift compiler has.

Also included is a basic AST printer that is loosely modeled after `clang++ -ast-dump`. We use it here for unit testing.

Reviewed By: vitaut

Differential Revision: D61077276

fbshipit-source-id: 9b02483bc983a699a69210185a4277750821f06b