Implement whisker::print_ast
Summary:
This diff adds `whisker::print_ast` whose job is to add print a `whisker::ast` hierarchy to human readable form.

The `tree_printer` part has been split off from the AST traversal because we want to re-use that to print the `whisker::object` (context) hierarchy in a future diff.

Reviewed By: vitaut

Differential Revision: D61355029

fbshipit-source-id: 2f441c3863af1aad0be8e7f2d1d538b5e2fab193