Remove apache::thrift::Client<Foo>::header_ methods by default
Summary: This enforces the contract outlined in {D60782227}. In short, we're going to stop generating `AsyncClient::header_*` variants of methods. We've already added annotations for existing code that use the `header_*` methods. As long as the builds don't break, this is deleting unused code. This should help us reclaim some binary size headroom so we can implement `ClientInterceptor`s.

Reviewed By: nickbrekhus

Differential Revision: D60793177

fbshipit-source-id: c02254485b4f22b5215a2b26453f9365fb4f7829