Implement to_string(whisker::object)
Summary:
This diff implements `whisker::to_string(whisker::object)` which uses `tree_printer.h` added in {D61355029} that is already used to print the AST. This functionality is super helpful for debugging and outputting diagnostics.

All whisker data types are supported except for `native_object`. At present, the printing `native_object` just prints out `<native_object>` rather than recursing into it. In the future, as we build out the evaluation engine, the plan is to add a customization point in `native_object` that will allow it to populating a node within the output tree. This will, for example, just us output `mstch::object` hierarchy intermixed with `whisker::object` types.

Reviewed By: vitaut

Differential Revision: D61328774

fbshipit-source-id: 905558e5112e74e0191e69b136de39dd308184be