Add two ResourcePoolSet::setResourcePool API variants.
Summary:
This adds two new variants to the `ResourcePoolSet::setResourcePool` API call.

1. A variant that takes none of `requestPile`, `executor`, and `concurrencyController`.
2. A variant that takes neither `requestPile` nor `concurrencyController`.

Note: The original variant API is unchanged.
Note: The two new variants do not include a `priorityHint` parameter as it was marked deprecated in the original variant API.

The two new variants require that none of the arguments are nullptr. The behavior of the original API variant is preserved (it allows some configurations of `requestPile`, `executor`, and `concurrencyController` to be nullptr).

The API changes in `ResourcePoolSet` are reflected in the constructors of `ResourcePool` (similar two new variants show up there).

There were only a few call-sites within our thrift code for these APIs, so those are included in this diff to show that they work.

Reviewed By: robertroeser

Differential Revision: D61629645

fbshipit-source-id: 7cbfebc87a65c1a0fb7fdd9f0da42458d49ee8cb