Check that string literals are valid UTF-8
Summary:
Check that the source text in string literals is valid UTF-8 as documented in https://github.com/facebook/fbthrift/blob/main/thrift/doc/idl/index.md#lexical-elements. We don't need to check other tokens because invalid code units already result in lexer errors. In principle we should also check comments but it is less important and can be done separately.

Strictly speaking this is a breaking change but most code generators such as Python and Java already require valid Unicode so it's unlikely to cause a problem in practice. In the worst case it will result in a compilation error, catching nonportable Thrift code.

Reviewed By: yoney

Differential Revision: D61694803

fbshipit-source-id: 92b07f5d0727126c066046459c6b2cb32ab64576